// src/app/models/cakeRequest.ts
export type CakeRequest = {
  id?: number;
  deliveryDate?: Date | string;
  customerName?: string;
  email?: string;
  phone?: number;
  pincode?: number;
  houseNumber?: number;
  city?: string;
  state?: string;
  quantity?: number;
  price?: number;
  totalPrice?: number;
  cakeName?: string;
};